Potential fix for zOS failures
Current theory is that when `IIHF` is enabled we pollute `enumReg` in
`genInitObjectHeader(...)`. By the time we get to updating `dataAddr`
slot `enumReg` contains both the array class and the size. This is what
was likely resulting in a 0 size variable length array being handled
as contiguous instead of discontiguous. Using `enumCopyReg` should
address this issue.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>