Set symbol reference as internal pointer
Now that we are loading the address of the first array element we don't
need to set the node as internal pointer. Instead, we can set the symbol
reference of the `aload` node as internal pointer, and use the symbol
reference of the array base node as the pinning array pointer.