Update isDiscontiguousArray(...) query
If off heap allocation is enabled this query should always return false.

Update isDiscontiguousArray(...) call sites

Update the call sites to also check if off heap allocation is enabled.
This is needed at some call sites to determine if dataAddr pointer
needs to be loaded to access data portion of the array.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>