Update getAddressOfElement to use dataAddr pointer
Call sites of `getAddressOfElement(...)` must be updated to work better
with off heap.

Instead of expecting the caller to add array header size to offset, it's
now done in the function. Header size is added to the offset only if
the array is contiguous and off heap allocation is disabled.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>