Use 32bit compare for array size in inline array allocation sequence
We need to use 32 bit compare because the top half of the size reg can contain
class pointer. On z, Class pointer is written to top half by IIHF in
genInitObjectHeader. I have updated the compare instruction on x as well to
avoid future issues related to garbage in top half of the size register.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>